Extract MarkdownContent from DocExplorer
